# perf(core): skip unnecessary `access` call in `fetchPackageFromCache`