--- a/packages/yarnpkg-core/sources/Project.ts
+++ b/packages/yarnpkg-core/sources/Project.ts
@@ -1884,8 +1884,6 @@ export class Project {
           : `${lastEntryRemoved} appeared to be unused and was removed`,
       );
     }
-
-    cache.markedFiles.clear();
   }
 }